Moved apps into presentation
