+++ src/presentation/apps/popup/index.html
@@ -9,6 +9,6 @@
 
   <body>
     <div id="root"></div>
-    <script type="module" src="/sidebar/main.tsx"></script>
+    <script type="module" src="main.tsx"></script>
   </body>
 </html>
