+++ src/presentation/apps/sidebar/App.tsx
@@ -7,36 +7,36 @@ import { SendersContainer } from "./components/sendersContainer.tsx";
 import { DeclutterHeader } from "./components/header.tsx";
 // import { useActions } from "../_shared/providers/actionsContext.tsx";
 import { ModalProvider } from "./providers/modalContext.tsx";
-import { ThemeContext } from "../_shared/providers/themeContext.ts";
+import { ThemeContext } from "../../../_shared/providers/themeContext.ts";
 import ThemeToggle from "./components/themeToggle.tsx";
-import { AppProvider } from "../presentation/providers/app_provider.tsx";
+import { AppProvider } from "../../providers/app_provider.tsx";
 
 function App() {
   const { theme } = useContext(ThemeContext);
 
   return (
     <AppProvider>
-        <ModalProvider>
-          <div id="declutter-body" className={theme}>
-            <DeclutterHeader />
+      <ModalProvider>
+        <div id="declutter-body" className={theme}>
+          <DeclutterHeader />
 
-            <div className="button-bar">
-              <div className="sender-actions">
-                <ActionButton id="unsubscribe-button" />
-                <ActionButton id="delete-button" />
-              </div>
-
-              <div style={{ display: "flex" }}>
-                <ReloadButton />
-                <ThemeToggle />
-              </div>
+          <div className="button-bar">
+            <div className="sender-actions">
+              <ActionButton id="unsubscribe-button" />
+              <ActionButton id="delete-button" />
             </div>
 
-            <SendersContainer />
-
-            <ModalPopup />
+            <div style={{ display: "flex" }}>
+              <ReloadButton />
+              <ThemeToggle />
+            </div>
           </div>
-        </ModalProvider>
+
+          <SendersContainer />
+
+          <ModalPopup />
+        </div>
+      </ModalProvider>
     </AppProvider>
   );
 }
