+++ src/presentation/apps/sidebar/components/loadingBar.css
