+++ src/presentation/apps/sidebar/components/reloadButton.css
