+++ src/presentation/apps/sidebar/components/senderLine.css
