+++ src/presentation/apps/sidebar/components/senderLineSkeleton.tsx
