+++ src/presentation/apps/sidebar/components/themeToggle.css
