+++ src/presentation/apps/sidebar/components/toggle.css
