+++ src/presentation/apps/sidebar/components/toggleSwitch.tsx
