+++ src/presentation/apps/sidebar/index.html
@@ -4,12 +4,11 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image" href="/images/icon-48.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/tutorial/index.css" />
     <title>InboxWhiz</title>
   </head>
 
   <body>
     <div id="root"></div>
-    <script type="module" src="/tutorial/main.tsx"></script>
+    <script type="module" src="main.tsx"></script>
   </body>
 </html>
