+++ src/presentation/apps/sidebar/main.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "../index.css";
+import "../../../index.css";
 import App from "./App.tsx";
 import { AllGlobalProviders } from "./providers/allGlobalProviders.tsx";
