+++ src/presentation/apps/sidebar/providers/allGlobalProviders.tsx
@@ -1,0 +1,12 @@
+import { ActionsProvider } from "../../../../_shared/providers/actionsContext";
+import { ThemeProvider } from "../../../../_shared/providers/ThemeProvider";
+
+export const AllGlobalProviders: React.FC<{ children: React.ReactNode }> = ({
+  children,
+}) => {
+  return (
+    <ThemeProvider>
+      <ActionsProvider>{children}</ActionsProvider>
+    </ThemeProvider>
+  );
+};
