+++ src/presentation/apps/tutorial/assets/top-senders.gif
