+++ src/presentation/apps/tutorial/assets/unsubscribe.gif
