+++ src/presentation/apps/tutorial/components/successIcon.tsx
