+++ src/presentation/apps/tutorial/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import "../index.css";
+import "../../../index.css";
 import Tutorial from "./Tutorial.tsx";
-import { ThemeProvider } from "../_shared/providers/ThemeProvider.tsx";
+import { ThemeProvider } from "../../providers/ThemeProvider.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
